+++ bld/wgml1/c/gsflngth.c
@@ -32,7 +32,7 @@
 /*                                                                         */
 /***************************************************************************/
 
-condcode    scr_length( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_length( parm parms[MAX_FUN_PARMS], int parmcount, char **result, int32_t ressize )
 {
     tok_type        parm1;
     int             len;
